Remove jtl-loader as it's overshadowed by external-data-loader
