+++ bzt/modules/external.py
@@ -3,11 +3,10 @@ import re
 from bzt import TaurusInternalException, TaurusConfigError
 from bzt.engine import ScenarioExecutor
 from bzt.modules.ab import TSVDataReader
-from bzt.modules.aggregator import AggregatorListener
-from bzt.modules.aggregator import ConsolidatingAggregator, ResultsReader, DataPoint
+from bzt.modules.aggregator import AggregatorListener, ConsolidatingAggregator, DataPoint
 from bzt.modules.gatling import DataLogReader as GatlingLogReader
 from bzt.modules.grinder import DataLogReader as GrinderLogReader
-from bzt.modules.jmeter import JTLReader, JTLErrorsReader
+from bzt.modules.jmeter import JTLReader, XMLJTLReader
 from bzt.modules.pbench import PBenchKPIReader
 
 
@@ -67,32 +66,3 @@ class ExternalReportAnalyzer(ScenarioExecutor, AggregatorListener):
             return True
         else:
             self._prev_ts = self._last_ts
-
-
-class XMLJTLReader(JTLErrorsReader, ResultsReader):
-    def __init__(self, filename, parent_logger):
-        super(XMLJTLReader, self).__init__(filename, parent_logger)
-        self.items = []
-
-    def _read(self, final_pass=False):
-        self.read_file()
-        while self.items:
-            yield self.items.pop(0)
-
-    def _parse_element(self, elem):
-        tstmp = int(int(elem.get("ts")) / 1000)
-        label = elem.get("lb")
-        rtm = int(elem.get("t")) / 1000.0
-        ltc = int(elem.get("lt")) / 1000.0 if "lt" in elem.attrib else 0
-        cnn = int(elem.get("ct")) / 1000.0 if "ct" in elem.attrib else 0
-        byte_count = int(elem.get("by")) if "by" in elem.attrib else 0
-        concur = int(elem.get("na")) if "na" in elem.attrib else 0
-        trname = ''
-
-        rcd = elem.get("rc")
-        message = self.get_failure_message(elem)
-        if message is None:
-            message = elem.get('rm')
-
-        error = message if elem.get("s") == "false" else None
-        self.items.append((tstmp, label, concur, rtm, cnn, ltc, rcd, error, trname, byte_count))
