--- site/dat/docs/changes/add-jtl-loader-executor.change
@@ -1,1 +1,0 @@
-Add `jtl-loader` executor specifically for loading and processing existing JTL files
